# refactor: remove `Clone` bound for storage key and value