--- a/foyer-common/Cargo.toml
+++ b/foyer-common/Cargo.toml
@@ -14,11 +14,13 @@ readme = "../README.md"
 normal = ["foyer-workspace-hack"]
 
 [dependencies]
+ahash = "0.8"
 anyhow = "1.0"
 bytes = "1"
 cfg-if = "1"
 crossbeam = "0.8"
 foyer-workspace-hack = { version = "0.4", path = "../foyer-workspace-hack" }
+hashbrown = "0.14"
 itertools = "0.12"
 nix = { version = "0.28", features = ["fs"] }
 parking_lot = { version = "0.12", features = ["arc_lock"] }