--- a/foyer-common/src/code.rs
+++ b/foyer-common/src/code.rs
@@ -22,16 +22,10 @@ pub trait Value: Send + Sync + 'static {}
 impl<T: Send + Sync + 'static + std::hash::Hash + Eq> Key for T {}
 impl<T: Send + Sync + 'static> Value for T {}
 
-/// [`StorageKey`] is required to implement [`Clone`].
-///
-/// If cloning a [`StorageKey`] is expensive, wrap it with [`Arc`].
 // TODO(MrCroxx): use `expect` after `lint_reasons` is stable.
-pub trait StorageKey: Key + Debug + Clone + Serialize + DeserializeOwned {}
-impl<T> StorageKey for T where T: Key + Debug + Clone + Serialize + DeserializeOwned {}
+pub trait StorageKey: Key + Debug + Serialize + DeserializeOwned {}
+impl<T> StorageKey for T where T: Key + Debug + Serialize + DeserializeOwned {}
 
-/// [`StorageValue`] is required to implement [`Clone`].
-///
-/// If cloning a [`StorageValue`] is expensive, wrap it with [`Arc`].
 // TODO(MrCroxx): use `expect` after `lint_reasons` is stable.
 pub trait StorageValue: Value + 'static + Debug + Clone + Serialize + DeserializeOwned {}
 impl<T> StorageValue for T where T: Value + Debug + Clone + Serialize + DeserializeOwned {}