--- a/foyer-common/src/lib.rs
+++ b/foyer-common/src/lib.rs
@@ -14,6 +14,7 @@
 
 #![cfg_attr(coverage_nightly, feature(coverage_attribute))]
 
+pub mod arc_key_hash_map;
 pub mod async_queue;
 pub mod batch;
 pub mod bits;