--- a/foyer-storage/src/judge.rs
+++ b/foyer-storage/src/judge.rs
@@ -43,6 +43,8 @@ impl Judges {
         }
     }
 
+    // TODO(MrCroxx): use `expect` after `lint_reasons` is stable.
+    #[allow(unused)]
     pub fn get(&mut self, index: usize) -> bool {
         self.judge.get(index)
     }