--- a/foyer-storage/src/reinsertion/rated_ticket.rs
+++ b/foyer-storage/src/reinsertion/rated_ticket.rs
@@ -81,8 +81,4 @@ where
 
         res
     }
-
-    fn on_insert(&self, _key: &Self::Key, _judge: bool) {}
-
-    fn on_drop(&self, _key: &Self::Key, _judge: bool) {}
 }