## What's changed and what's your intention?

> Please explain **IN DETAIL** what the changes are in this PR and why they are needed. :D

As title.

Modification:

- Use `Arc<K>` and `Arc<V>` instead of `K` and `V` if necessary.
- Remove `on_insert` and `on_drop` callback for admission and reinsertion policies.
- Lookup return `Option<CachedEntry<K, V>>` which holds shared pointer or owned pointer instead of `Option<V>` directly.
- Catalog uses `ArcKeyHashMap` with `hashbrown::HashTable` to implement key comparing (which doesn't require `Arc<K>: Borrow<Q>`) instead of std hash map.

## Checklist

- [x] I have written the necessary rustdoc comments
- [x] I have added the necessary unit tests and integration tests
- [x] I have passed `make all` (or `make fast` instead if the old tests are not modified) in my local environment.

## Related issues or PRs (optional)
close #364