@MrCroxx 2024-04-19
## Benchmark

### cmd

```bach
rm -rf /p44pro/foyer && cargo build --release && RUST_BACKTRACE=1 RUST_LOG=warn ./target/release/foyer-storage-bench --dir /p44pro/foyer --capacity 204800 --file-size 64 --lookup-range 10000 --flushers 12 --reclaimers 12 --time 180 --writers 64 --w-rate 8 --ticket-insert-rate-limit 500 --readers 1024 --r-rate 1 --runtime --metrics --compression none --io-size 262144 --distribution zipf --distribution-zipf-s 0.8
```

### main (as https://github.com/MrCroxx/foyer/pull/366#issuecomment-2063651910)

```
Total:
disk total iops: 3751.2
disk total throughput: 478.7 MiB/s
disk read iops: 288.1
disk read throughput: 19.1 MiB/s
disk write iops: 3463.0
disk write throughput: 459.6 MiB/s
insert iops: 6919.8/s
insert throughput: 432.5 MiB/s
insert lat p50: 35us
insert lat p90: 72us
insert lat p99: 515us
insert lat p999: 1159us
insert lat p9999: 2303us
insert lat p99999: 6783us
insert lat pmax: 8383us
get iops: 16359.1/s
get miss: 0.00%
get throughput: 1.0 GiB/s
get hit lat p50: 1087us
get hit lat p90: 2351us
get hit lat p99: 3375us
get hit lat p999: 5727us
get hit lat p9999: 13439us
get hit lat p99999: 23679us
get hit lat pmax: 24191us
get miss lat p50: 0us
get miss lat p90: 0us
get miss lat p99: 0us
get miss lat p999: 0us
get miss lat p9999: 0us
get miss lat p99999: 0us
get miss lat pmax: 0us
```

### PR

```
Total:
disk total iops: 4247.5
disk total throughput: 511.8 MiB/s
disk read iops: 783.1
disk read throughput: 52.0 MiB/s
disk write iops: 3464.3
disk write throughput: 459.8 MiB/s
insert iops: 6922.4/s
insert throughput: 432.7 MiB/s
insert lat p50: 36us
insert lat p90: 76us
insert lat p99: 327us
insert lat p999: 847us
insert lat p9999: 1479us
insert lat p99999: 2847us
insert lat pmax: 12095us
get iops: 19723.8/s
get miss: 16.95%
get throughput: 1.0 GiB/s
get hit lat p50: 931us
get hit lat p90: 1711us
get hit lat p99: 2623us
get hit lat p999: 6687us
get hit lat p9999: 11903us
get hit lat p99999: 20479us
get hit lat pmax: 24703us
get miss lat p50: 579us
get miss lat p90: 927us
get miss lat p99: 1375us
get miss lat p999: 4287us
get miss lat p9999: 24959us
get miss lat p99999: 25471us
get miss lat pmax: 25983us
```